Split storage.rb into storage/filesystem.rb and storage/s3.rb.
